feat: Add resolve command for review thread management
A new `resolve` command was introduced to enable marking GitHub review comment
threads as resolved or unresolved. This command accepts a `COMMENT_ID` and
an optional `PR_NUMBER`, defaulting to the current branch's pull request.
The `--unresolve` flag was added to allow unresolving a thread.
The `github.Client` package was updated with an `UnresolveThread` method
to perform the GraphQL mutation for unresolving threads.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>